refactor(front): move map submission components to subdir
