+++ apps/frontend/src/app/components/map-forms/map-image-selection/image-selection-item.class.ts
