+++ apps/frontend/src/app/components/map-forms/map-image-selection/map-image-selection.component.html
