+++ apps/frontend/src/app/components/map-forms/map-leaderboard-selection/map-leaderboards-selection.component.html
